refactor(rust): 移除内联 runtime/llm_client，抽出 host_runtime 与 llm_types
- 删除 runtime.rs、llm_client.rs；ports 移除 LlmTransport、Telemetry 等仅旧内联路径使用的 trait 与类型

- 新增 host_runtime.rs（RuntimeEvent、工具请求 UI 展示）；llm_types.rs（LlmMessage）

- 精简 adapters；调整 session、ts_bridge、tui、file_reference；更新 Copilot 架构说明
