+++ spirit-agent/src/llm_types.rs
@@ -1,0 +1,6 @@
+#[derive(Clone, Debug)]
+pub struct LlmMessage {
+    pub role: &'static str,
+    pub content: String,
+    pub image_paths: Vec<String>,
+}
